[msm.ui]: moved MSMEstimator from estimation.obj to ui
