+++ pyemma/msm/api.py
@@ -31,7 +31,7 @@ __docformat__ = "restructuredtext en"
 
 from flux import tpt as tpt_factory
 from pyemma.msm.estimation.obj.dtraj_stats import DiscreteTrajectoryStats
-from pyemma.msm.estimation.obj.msm_estimator import MSMEstimator
+from pyemma.msm.ui.msm_estimator import MSMEstimator
 from ui import ImpliedTimescales
 from ui import MSM
 from ui import EstimatedMSM
