+++ pyemma/msm/ui/msm_estimator.py
